requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clickhouse-connect==0.7.7
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
load-dotenv==0.1.0
lz4==4.3.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
StrEnum==0.4.15
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1
zstandard==0.22.0
schedule==1.2.1
